Implement HTTP chunked encoder and decoder
When Content-Length field is missing from the HTTP request headers, warp
can now consume the body of the request by decoding the HTTP
chunked-encoding.

An example of missing Content-Length appears when sending a stream of
data of unknown size to the server. Such a request can be generated by
the following command:

  echo Text of unknown size | curl -T- http://localhost:1234/upload